+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/StandaloneApiServer.scala
@@ -10,106 +10,84 @@ import com.daml.buildinfo.BuildInfo
 import com.daml.error.ErrorCodesVersionSwitcher
 import com.daml.ledger.api.auth.interceptor.AuthorizationInterceptor
 import com.daml.ledger.api.auth.{AuthService, Authorizer}
-import com.daml.ledger.api.domain
 import com.daml.ledger.api.health.HealthChecks
 import com.daml.ledger.configuration.LedgerId
 import com.daml.ledger.participant.state.{v2 => state}
-import com.daml.ledger.resources.{Resource, ResourceContext, ResourceOwner}
+import com.daml.ledger.resources.ResourceOwner
 import com.daml.lf.data.Ref
-import com.daml.lf.data.Time.Timestamp
-import com.daml.lf.engine.{Engine, ValueEnricher}
+import com.daml.lf.engine.Engine
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.metrics.Metrics
 import com.daml.platform.configuration.{
   CommandConfiguration,
   PartyConfiguration,
-  ServerRole,
   SubmissionConfiguration,
 }
-import com.daml.platform.index.JdbcIndex
-import com.daml.platform.packages.InMemoryPackageStore
 import com.daml.platform.services.time.TimeProviderType
-import com.daml.platform.store.LfValueTranslationCache
 import com.daml.ports.{Port, PortFiles}
 import io.grpc.{BindableService, ServerInterceptor}
 import scalaz.{-\/, \/-}
-import java.io.File
 import java.time.Clock
 
+import com.daml.ledger.participant.state.index.v2.IndexService
 import com.daml.telemetry.TelemetryContext
 
 import scala.collection.immutable
 import scala.concurrent.ExecutionContextExecutor
 import scala.util.{Failure, Success, Try}
 
-final class StandaloneApiServer(
-    ledgerId: LedgerId,
-    config: ApiServerConfig,
-    commandConfig: CommandConfiguration,
-    partyConfig: PartyConfiguration,
-    submissionConfig: SubmissionConfiguration,
-    optWriteService: Option[state.WriteService],
-    authService: AuthService,
-    healthChecks: HealthChecks,
-    metrics: Metrics,
-    timeServiceBackend: Option[TimeServiceBackend] = None,
-    otherServices: immutable.Seq[BindableService] = immutable.Seq.empty,
-    otherInterceptors: List[ServerInterceptor] = List.empty,
-    engine: Engine,
-    servicesExecutionContext: ExecutionContextExecutor,
-    lfValueTranslationCache: LfValueTranslationCache.Cache,
-    checkOverloaded: TelemetryContext => Option[state.SubmissionResult] =
-      _ => None, // Used for Canton rate-limiting
-)(implicit actorSystem: ActorSystem, materializer: Materializer, loggingContext: LoggingContext)
-    extends ResourceOwner[ApiServer] {
-
+object StandaloneApiServer {
   private val logger = ContextualizedLogger.get(this.getClass)
 
-  // Name of this participant,
-  val participantId: Ref.ParticipantId = config.participantId
+  def apply(
+      indexService: IndexService,
+      ledgerId: LedgerId,
+      config: ApiServerConfig,
+      commandConfig: CommandConfiguration,
+      partyConfig: PartyConfiguration,
+      submissionConfig: SubmissionConfiguration,
+      optWriteService: Option[state.WriteService],
+      authService: AuthService,
+      healthChecks: HealthChecks,
+      metrics: Metrics,
+      timeServiceBackend: Option[TimeServiceBackend] = None,
+      otherServices: immutable.Seq[BindableService] = immutable.Seq.empty,
+      otherInterceptors: List[ServerInterceptor] = List.empty,
+      engine: Engine,
+      servicesExecutionContext: ExecutionContextExecutor,
+      checkOverloaded: TelemetryContext => Option[state.SubmissionResult] =
+        _ => None, // Used for Canton rate-limiting
+  )(implicit
+      actorSystem: ActorSystem,
+      materializer: Materializer,
+      loggingContext: LoggingContext,
+  ): ResourceOwner[ApiServer] = {
+    val participantId: Ref.ParticipantId = config.participantId
 
-  override def acquire()(implicit context: ResourceContext): Resource[ApiServer] = {
-    val packageStore = loadDamlPackages()
-    preloadPackages(packageStore)
+    def writePortFile(port: Port): Try[Unit] = {
+      config.portFile match {
+        case Some(path) =>
+          PortFiles.write(path, port) match {
+            case -\/(err) => Failure(new RuntimeException(err.toString))
+            case \/-(()) => Success(())
+          }
+        case None =>
+          Success(())
+      }
+    }
 
-    val valueEnricher = new ValueEnricher(engine)
+    val errorCodesVersionSwitcher = new ErrorCodesVersionSwitcher(
+      config.enableSelfServiceErrorCodes
+    )
+    val authorizer = new Authorizer(
+      Clock.systemUTC.instant _,
+      ledgerId,
+      participantId,
+      errorCodesVersionSwitcher,
+    )
+    val healthChecksWithIndexService = healthChecks + ("index" -> indexService)
 
-    val owner = for {
-      indexService <- JdbcIndex
-        .owner(
-          serverRole = ServerRole.ApiServer,
-          ledgerId = domain.LedgerId(ledgerId),
-          participantId = participantId,
-          jdbcUrl = config.jdbcUrl,
-          databaseConnectionPoolSize = config.databaseConnectionPoolSize,
-          databaseConnectionTimeout = config.databaseConnectionTimeout,
-          eventsPageSize = config.eventsPageSize,
-          eventsProcessingParallelism = config.eventsProcessingParallelism,
-          acsIdPageSize = config.acsIdPageSize,
-          acsIdFetchingParallelism = config.acsIdFetchingParallelism,
-          acsContractFetchingParallelism = config.acsContractFetchingParallelism,
-          servicesExecutionContext = servicesExecutionContext,
-          metrics = metrics,
-          lfValueTranslationCache = lfValueTranslationCache,
-          enricher = valueEnricher,
-          maxContractStateCacheSize = config.maxContractStateCacheSize,
-          maxContractKeyStateCacheSize = config.maxContractKeyStateCacheSize,
-          enableMutableContractStateCache = config.enableMutableContractStateCache,
-          maxTransactionsInMemoryFanOutBufferSize = config.maxTransactionsInMemoryFanOutBufferSize,
-          enableInMemoryFanOutForLedgerApi = config.enableInMemoryFanOutForLedgerApi,
-          enableSelfServiceErrorCodes = config.enableSelfServiceErrorCodes,
-        )
-        .map(index => new SpannedIndexService(new TimedIndexService(index, metrics)))
-      errorCodesVersionSwitcher = new ErrorCodesVersionSwitcher(
-        config.enableSelfServiceErrorCodes
-      )
-      authorizer = new Authorizer(
-        Clock.systemUTC.instant _,
-        ledgerId,
-        participantId,
-        errorCodesVersionSwitcher,
-      )
-      healthChecksWithIndexService = healthChecks + ("index" -> indexService)
+    for {
       executionSequencerFactory <- new ExecutionSequencerFactoryOwner()
       apiServicesOwner = new ApiServices.Owner(
         participantId = participantId,
@@ -145,7 +123,7 @@ final class StandaloneApiServer(
         config.tlsConfig,
         AuthorizationInterceptor(
           authService,
-          executionContext,
+          servicesExecutionContext,
           errorCodesVersionSwitcher,
         ) :: otherInterceptors,
         servicesExecutionContext,
@@ -157,50 +135,6 @@ final class StandaloneApiServer(
         s"Initialized API server version ${BuildInfo.Version} with ledger-id = $ledgerId, port = ${apiServer.port}, dar file = ${config.archiveFiles}"
       )
       apiServer
-    }
-
-    owner.acquire()
-  }
-
-  private def preloadPackages(packageContainer: InMemoryPackageStore): Unit = {
-    for {
-      (pkgId, _) <- packageContainer.listLfPackagesSync()
-      pkg <- packageContainer.getLfPackageSync(pkgId)
-    } {
-      engine
-        .preloadPackage(pkgId, pkg)
-        .consume(
-          { _ =>
-            sys.error("Unexpected request of contract")
-          },
-          packageContainer.getLfPackageSync,
-          { _ =>
-            sys.error("Unexpected request of contract key")
-          },
-        )
-      ()
-    }
-  }
-
-  private def loadDamlPackages(): InMemoryPackageStore = {
-    // TODO is it sensible to have all the initial packages to be known since the epoch?
-    config.archiveFiles
-      .foldLeft[Either[(String, File), InMemoryPackageStore]](Right(InMemoryPackageStore.empty)) {
-        case (storeE, f) =>
-          storeE.flatMap(_.withDarFile(Timestamp.now(), None, f).left.map(_ -> f))
-      }
-      .fold({ case (err, file) => sys.error(s"Could not load package $file: $err") }, identity)
-  }
-
-  private def writePortFile(port: Port): Try[Unit] = {
-    config.portFile match {
-      case Some(path) =>
-        PortFiles.write(path, port) match {
-          case -\/(err) => Failure(new RuntimeException(err.toString))
-          case \/-(()) => Success(())
-        }
-      case None =>
-        Success(())
     }
   }
 }
