+++ ledger/participant-state/kvutils/app/BUILD.bazel
@@ -46,6 +46,7 @@ da_scala_library(
         "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-state",
+        "//ledger/participant-state-index",
         "//ledger/participant-state-metrics",
         "//ledger/participant-state/kvutils",
         "//libs-scala/contextualized-logging",
